+++ .env.example
@@ -1,0 +1,6 @@
+DB_USER=HelloTalk
+DB_PASSWORD=Xuon587WQdPMuqZF
+ACCESS_TOKEN=50000ce76960446240ccafc854e56e777e33634ab46f1541fd06334b277d4b68ce2a7a4f9acf6061db559b11e143ab84a53563d3936c33118d9c7bb27c26977c
+APP_NAME="LEAD EDUCARE"
+DATABASE=mongodb://127.0.0.1:27017/lead-educare
+PORT=8000
